Split Platform and Collector out of elf.rs.
